Fix linux-x86_64 build.
Fix various 64-bitness issues in the source code to
make the --try-64 option work again on Linux. Note that
the generated binary is not faster than its 32-bit variant
when it comes to benchmarking the boot sequence.

+++ android/utils/timezone.c
@@ -185,7 +185,7 @@ compare_timezone_to_localtime( ScanDataRec*  scan,
     }
 
     if ( st.st_size != scan->localtime_st.st_size ) {
-        D( " size mistmatch (%lld != %lld)\n", st.st_size, scan->localtime_st.st_size );
+        D( " size mistmatch (%zd != %zd)\n", (size_t)st.st_size, (size_t)scan->localtime_st.st_size );
         return 0;
     }
